Find an element in a BST from its given rank
Explanation of the Code:
Node Class:

The Node class represents a node in the BST. It has an integer data, pointers to the left and right children, and a count of nodes in the subtree rooted at this node.
insertintoBST Function:

This function inserts a node with given data d into the BST and updates the count of nodes in the subtree.
If the tree is empty, it creates a new node.
If the data to be inserted is greater than the root's data, it inserts in the right subtree.
If the data to be inserted is less than or equal to the root's data, it inserts in the left subtree.
The count of nodes in the subtree rooted at the current node is incremented.
countNodes Function:

This function counts the total number of nodes in the subtree rooted at root by recursively counting nodes in the left and right subtrees.
elementofrankx Function:

This function finds the node with the given rank x.
It uses a while loop to traverse the tree. If the right subtree is not NULL, it checks if the rank of the current node matches x.
If x is less than the rank of the current node, it moves to the right subtree.
Otherwise, it adjusts x and moves to the left subtree.
If the right subtree is NULL, it checks if x is the rank of the current node and adjusts x to move to the left subtree if needed.
takeInput Function:

This function takes input from the user to create the BST. It reads integers from the input until -1 is entered. For each input, it calls insertintoBST to insert the data into the BST.
main Function:

The entry point of the program. It initializes the root of the BST to NULL and prompts the user to enter data to create the BST.
It then prompts the user to enter the rank and finds the element with the given rank using elementofrankx.
If the element is found, it prints the data. Otherwise, it prints "Rank out of bounds".